# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LUSIRIUS TEST Documentation",
        "description": "API Documentation",
        "contact": {
            "email": "viryl1515@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get list of reports",
                "description": "Returns list of reports",
                "operationId": "getReportsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/report-image": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create new report",
                "description": "Create new report",
                "operationId": "createReport",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "The image to be reported (the link or byte)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "callback",
                        "in": "query",
                        "description": "The Callback endpoint use to send report result when it's available",
                        "required": false,
                        "schema": {
                            "type": "url"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageReport"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/reevaluate-report/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Revaluate existing report",
                "description": "Revaluate existing report",
                "operationId": "revaluateReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageReport"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/approve-report/{id}": {
            "put": {
                "tags": [
                    "Reports"
                ],
                "summary": "Approve report",
                "description": "Approve report",
                "operationId": "approveReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "approve",
                        "in": "query",
                        "description": "approval value",
                        "required": true,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageReport"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/update-report-callback/{id}": {
            "put": {
                "tags": [
                    "Reports"
                ],
                "summary": "Update callback endpoint",
                "description": "Update callback endpoint. Hit this to update the image report callback endpoint.",
                "operationId": "updateCallbackReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "callback",
                        "in": "query",
                        "description": "callback link value",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "http://localhost/api/callback-test"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageReport"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/callback/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Callback report. Hit this to get the image report moderation by providing the report id.",
                "description": "Callback report. Hit this to get the image report moderation by providing the report id.",
                "operationId": "callbackReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportCallbackResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/callback-test": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Callback report test",
                "description": "Callback report test",
                "operationId": "callbackTester",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReportCallbackResource": {
                "title": "ReportCallbackResource",
                "description": "Report callback resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImageCallbackReport"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ReportCallbackResource"
                }
            },
            "ReportResource": {
                "title": "ReportResource",
                "description": "Report resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImageReport"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ReportResource"
                }
            },
...
        }
    },
...
}